+++ CHANGELOG.md
@@ -45,3 +45,4 @@
 - [2026-04-02 09:27] feat(network): integrate custom telemetry logger in network
 - [2026-04-02 09:38] docs(theme): update readme and API contracts for theme
 - [2026-04-02 09:59] feat(ui): add bi-directional event bus handler for ui
+- [2026-04-02 10:09] feat(theme): add async theme stream provider